+++ packages/styleguide/.storybook/decorators/theme.tsx
@@ -1,11 +1,16 @@
 import React from 'react';
-import { ThemeContext } from '@emotion/react';
-import { theme } from '@codecademy/gamut-styles';
+import { CacheProvider, ThemeContext } from '@emotion/react';
 
-export const withTheme = (Story) => {
-  return (
-    <ThemeContext.Provider value={theme}>
-      <Story />
-    </ThemeContext.Provider>
-  );
-};
+import { theme, createEmotionCache } from '@codecademy/gamut-styles';
+
+const cache = createEmotionCache();
+
+/**
+ * Story functions must be called as a regular function to avoid full-remounts
+ * See: https://github.com/storybookjs/storybook/issues/12255
+ */
+export const withEmotion = (Story: any) => (
+  <CacheProvider value={cache}>
+    <ThemeContext.Provider value={theme}>{Story()}</ThemeContext.Provider>
+  </CacheProvider>
+);